+++ .changes/v2.25.0/679-features.md
@@ -3,5 +3,6 @@
 * Added `VCDClient` methods `GetAllSolutionAddonInstanceByName`, `GetAllSolutionAddonInstances`,
   `GetSolutionAddOnInstanceById` [GH-679]
 * Added `SolutionAddOn` methods `CreateSolutionAddOnInstance`, `GetAllInstances`,
-  `GetInstanceByName` [GH-679]
-* Added `SolutionAddOnInstance` methods `Delete`, `RdeId`, `Publishing` [GH-679]
+  `GetInstanceByName`, `ValidateInputs`, `ConvertInputTypes` [GH-679]
+* Added `SolutionAddOnInstance` methods `GetParentSolutionAddOn`, `ReadCreationInputValues`
+  `Delete`, `RdeId`, `Publishing` [GH-679]
