+++ .changes/v2.25.0/679-improvements.md
@@ -1,0 +1,2 @@
+* Added convenience method `DefinedEntity.State()` that will automatically check if path to *State
+  has no nil pointers [GH-679]
