+++ govcd/solution_add_on_test.go
