+++ govcd/solution_addon_instance_test.go
